+++ src/ToolMate.php
@@ -4,12 +4,17 @@ namespace vaersaagod\toolmate;
 
 use Craft;
 use craft\base\Plugin;
+use craft\events\TemplateEvent;
 use craft\log\FileTarget;
+use craft\web\Application;
 use craft\web\twig\variables\CraftVariable;
 
+use craft\web\View;
+use vaersaagod\toolmate\services\CspService;
 use vaersaagod\toolmate\services\EmbedService;
 use vaersaagod\toolmate\services\MinifyService;
 use vaersaagod\toolmate\services\ToolService;
+use vaersaagod\toolmate\twigextensions\CspTwigExtension;
 use vaersaagod\toolmate\variables\ToolMateVariable;
 use vaersaagod\toolmate\twigextensions\ToolMateTwigExtension;
 use vaersaagod\toolmate\models\Settings;
@@ -21,10 +26,11 @@ use yii\base\Event;
  * @package   PluginMate
  * @since     1.0.0
  *
- * @property  EmbedService $embed
- * @property  ToolService $tool
- * @property  Settings $settings
- * @method    Settings getSettings()
+ * @property CspService $csp
+ * @property EmbedService $embed
+ * @property ToolService $tool
+ * @property Settings $settings
+ * @method Settings getSettings()
  */
 class ToolMate extends Plugin
 {
@@ -55,6 +61,7 @@ class ToolMate extends Plugin
 
         // Register services
         $this->setComponents([
+            'csp' => CspService::class,
             'embed' => EmbedService::class,
             'tool' => ToolService::class,
             'minify' => MinifyService::class,
@@ -74,12 +81,19 @@ class ToolMate extends Plugin
 
         // Add in our Twig extensions
         Craft::$app->view->registerTwigExtension(new ToolMateTwigExtension());
+        Craft::$app->view->registerTwigExtension(new CspTwigExtension());
 
         // Lets use our own log file
         Craft::getLogger()->dispatcher->targets[] = new FileTarget([
             'logFile' => '@storage/logs/toolmate.log',
             'categories' => ['vaersaagod\toolmate\*'],
         ]);
+
+        /**
+         * Maybe send a Content-Security-Policy header
+         */
+        $this->maybeSendCspHeader();
+
     }
 
     // Protected Methods
@@ -91,5 +105,48 @@ class ToolMate extends Plugin
     protected function createSettingsModel(): Settings
     {
         return new Settings();
+    }
+
+    protected function maybeSendCspHeader()
+    {
+
+        $cspConfig = ToolMate::getInstance()->getSettings()->csp;
+
+        if (!$cspConfig->enabled) {
+            return;
+        }
+
+        if (Craft::$app->getRequest()->getIsCpRequest() && !$cspConfig->enabledForCp) {
+            return;
+        }
+
+        // Replace hashed CSP nonces (this gets us around the template cache)
+        Event::on(
+            View::class,
+            View::EVENT_AFTER_RENDER_PAGE_TEMPLATE,
+            static function (TemplateEvent $event) {
+                \preg_match_all('/nonce="([^"]*)"/', $event->output, $matches);
+                $hashedNonces = $matches[1] ?? [];
+                for ($i = 0; $i < \count($hashedNonces); $i += 1) {
+                    if (!$unhashedNonce = Craft::$app->getSecurity()->validateData($hashedNonces[$i])) {
+                        continue;
+                    }
+                    [0 => $directive, 1 => $nonce] = \explode(':', $unhashedNonce);
+                    if (!ToolMate::getInstance()->csp->hasNonce($directive, $nonce)) {
+                        $nonce = ToolMate::getInstance()->csp->createNonce($directive);
+                    }
+                    $event->output = \str_replace($matches[0][$i], "nonce=\"$nonce\"", $event->output);
+                }
+            }
+        );
+
+        Event::on(
+            Application::class,
+            \yii\base\Application::EVENT_AFTER_REQUEST,
+            static function (Event $event) {
+                ToolMate::getInstance()->csp->setHeader();
+            }
+        );
+
     }
 }